+++ lib/src/rljson.dart
@@ -9,6 +9,9 @@ import 'package:gg_json_hash/gg_json_hash.dart';
 /// A simple json map
 typedef Rlmap = Map<String, dynamic>;
 
+/// A map of layers
+typedef Rllayers = Map<String, Rlmap>;
+
 /// Manages a normalized JSON data structure
 ///
 /// composed of layers '@layerA', '@layerB', etc.
@@ -25,7 +28,7 @@ class Rljson {
   final Rlmap originalData;
 
   /// Returns a map of layers containing a map of items for fast access
-  final Rlmap data;
+  final Rllayers data;
 
   // ...........................................................................
   /// Creates a new json containing the given data
@@ -54,7 +57,7 @@ class Rljson {
         // Layer does not exist yet. Insert all
         if (oldLayer == null) {
           mergedData[layer] = newLayer;
-          mergedMap[layer] = addedDataAsMap[layer];
+          mergedMap[layer] = addedDataAsMap[layer]!;
           continue;
         }
 
@@ -85,18 +88,98 @@ class Rljson {
   }
 
   // ...........................................................................
-  /// Allows to query data from the json
-  List<Rlmap> item({
-    required String layer,
-    required bool Function(Rlmap item) where,
-  }) {
-    final layerData = data[layer] as Map<String, Rlmap>?;
+  /// Returns the layer with the given name. Throws when name is not found.
+  Rllayers layer(String layer) {
+    final layerData = data[layer] as Rllayers?;
     if (layerData == null) {
       throw Exception('Layer not found: $layer');
     }
 
+    return layerData;
+  }
+
+  // ...........................................................................
+  /// Allows to query data from the json
+  List<Rlmap> items({
+    required String layer,
+    required bool Function(Rlmap item) where,
+  }) {
+    final layerData = this.layer(layer);
     final items = layerData.values.where(where).toList();
     return items;
+  }
+
+  // ...........................................................................
+  /// Allows to query data from the json
+  Rlmap item(
+    String layer,
+    String hash,
+  ) {
+    // Get layer
+    final layerData = data[layer];
+    if (layerData == null) {
+      throw Exception('Layer not found: $layer');
+    }
+
+    // Get item
+    final item = layerData[hash] as Rlmap?;
+    if (item == null) {
+      throw Exception(
+        'Item not found with hash "$hash" in layer "$layer"',
+      );
+    }
+
+    return item;
+  }
+
+  // ...........................................................................
+  /// Queries a value from data. Throws when layer or hash is not found.
+  dynamic value({
+    required String layer,
+    required String itemHash,
+    String? key,
+    String? key2,
+    String? key3,
+    String? key4,
+  }) {
+    // Get item
+    final resultItem = item(layer, itemHash);
+
+    // If no key is given, return the complete item
+    if (key == null) {
+      return resultItem;
+    }
+
+    // Get item value
+    final itemValue = resultItem[key];
+    if (itemValue == null) {
+      throw Exception(
+        'Key "$key" not found in item with hash "$itemHash" in layer "$layer"',
+      );
+    }
+
+    // Return item value when no link or links are not followed
+    if (!key.startsWith('@')) {
+      if (key2 != null) {
+        throw Exception('Invalid key "$key2". '
+            'Additional keys are only allowed for links. '
+            'But key "$key" points to a value.');
+      }
+
+      return itemValue;
+    }
+
+    // Follow links
+    final targetLayer = key;
+    final targetHash = itemValue as String;
+
+    return value(
+      layer: targetLayer,
+      itemHash: targetHash,
+      key: key2,
+      key2: key3,
+      key3: key4,
+    );
   }
 
   // ...........................................................................
@@ -105,7 +188,7 @@ class Rljson {
     final List<String> result = [];
     for (final layerEntry in data.entries) {
       final layer = layerEntry.key;
-      final layerData = layerEntry.value as Rlmap;
+      final layerData = layerEntry.value;
 
       for (final itemEntry in layerData.entries) {
         final item = itemEntry.value as Rlmap;
@@ -167,9 +250,11 @@ class Rljson {
       '_data': [
         {
           'keyA0': 'a0',
+          '_hash': 'KFQrf4mEz0UPmUaFHwH4T6',
         },
         {
           'keyA1': 'a1',
+          '_hash': 'YPw-pxhqaUOWRFGramr4B1',
         }
       ],
     },
@@ -204,9 +289,55 @@ class Rljson {
       '_data': [
         {
           '@layerA': 'KFQrf4mEz0UPmUaFHwH4T6',
+          '_hash': 'Cr5kvgDz5NGgbnuHu-Z05N',
         },
       ],
     },
+  });
+
+  // ...........................................................................
+  /// An example object
+  static final Rljson exampleWithDeepLink = Rljson.fromData({
+    '@a': {
+      '_data': [
+        {
+          '@b': 'Ji-ftHLbqQehV4aV0FlEO6',
+          'value': 'a',
+          '_hash': 'Q-oaM7xctsuFg_faf1lkhC',
+        }
+      ],
+      '_hash': 'GnSVp1CmoAo3rPiiGl44p-',
+    },
+    '@b': {
+      '_data': [
+        {
+          '@c': 'tlmxwmwJvFMyBYIYoA2k4K',
+          'value': 'b',
+          '_hash': 'ks2Zy5nlXx91deccdZtjvK',
+        }
+      ],
+      '_hash': 'uHEtXhILctvNH6zk6k4vDi',
+    },
+    '@c': {
+      '_data': [
+        {
+          '@d': '0tlKQklLrHoIhHZxdtcbmS',
+          'value': 'c',
+          '_hash': '3AzQ8kTQ-PjmW0FwY5mirx',
+        }
+      ],
+      '_hash': 'PE0bzvER5q3D-DgxswKzQM',
+    },
+    '@d': {
+      '_data': [
+        {
+          'value': 'd',
+          '_hash': '0tlKQklLrHoIhHZxdtcbmS',
+        },
+      ],
+      '_hash': 'WNq8zriiKUM_vn9DwrwAf0',
+    },
+    '_hash': '8_qxUEjOkDIq8Um7Z8OCt6',
   });
 
   // ######################
@@ -261,8 +392,8 @@ class Rljson {
   }
 
   // ...........................................................................
-  Rlmap _toMap(Rlmap data) {
-    final result = Rlmap();
+  Rllayers _toMap(Rlmap data) {
+    final result = <String, Rlmap>{};
 
     // Iterate all layers
     for (final layer in data.keys) {