+++ test/rljson_test.dart
@@ -49,9 +49,43 @@ void main() {
       });
     });
 
-    group('item(layer, where)', () {
+    group('layer(String layer)', () {
+      group('returns', () {
+        test('the layer when existing', () {
+          final layer = rljson.layer('@layerA');
+          expect(layer, {
+            'KFQrf4mEz0UPmUaFHwH4T6': {
+              'keyA0': 'a0',
+              '_hash': 'KFQrf4mEz0UPmUaFHwH4T6',
+            },
+            'YPw-pxhqaUOWRFGramr4B1': {
+              'keyA1': 'a1',
+              '_hash': 'YPw-pxhqaUOWRFGramr4B1',
+            },
+          });
+        });
+      });
+
+      group('throws', () {
+        test('when layer does not exist', () {
+          late final Exception exception;
+
+          try {
+            rljson.layer(
+              '@layerC',
+            );
+          } catch (e) {
+            exception = e as Exception;
+          }
+
+          expect(exception.toString(), 'Exception: Layer not found: @layerC');
+        });
+      });
+    });
+
+    group('items(layer, where)', () {
       test('returns the items that match the query', () {
-        final items = rljson.item(
+        final items = rljson.items(
           layer: '@layerA',
           where: (item) => item['keyA0'] == 'a0',
         );
@@ -60,21 +94,143 @@ void main() {
           {'keyA0': 'a0', '_hash': 'KFQrf4mEz0UPmUaFHwH4T6'},
         ]);
       });
+    });
+
+    group('item(layer, hash)', () {
+      group('returns', () {
+        test('the item when existing', () {
+          final item = rljson.item('@layerA', 'KFQrf4mEz0UPmUaFHwH4T6');
+          expect(item, {
+            'keyA0': 'a0',
+            '_hash': 'KFQrf4mEz0UPmUaFHwH4T6',
+          });
+        });
+      });
 
       group('throws', () {
-        test('when layer does not exist', () {
+        test('when layer is not available', () {
           late final Exception exception;
 
           try {
-            rljson.item(
-              layer: '@layerC',
-              where: (item) => item['keyA0'] == 'a0',
-            );
+            rljson.item('@layerC', 'KFQrf4mEz0UPmUaFHwH4T6');
           } catch (e) {
             exception = e as Exception;
           }
 
           expect(exception.toString(), 'Exception: Layer not found: @layerC');
+        });
+
+        test('when hash is not available', () {
+          late final Exception exception;
+
+          try {
+            rljson.item('@layerA', 'nonExistingHash');
+          } catch (e) {
+            exception = e as Exception;
+          }
+
+          expect(
+            exception.toString(),
+            'Exception: Item not found with hash "nonExistingHash" '
+            'in layer "@layerA"',
+          );
+        });
+      });
+    });
+
+    group('value(layer, hash, key, key2, key3, key4 followLinks)', () {
+      group('returns', () {
+        test('the value of the key of the item with hash in layer', () {
+          expect(
+            rljson.value(
+              layer: '@layerA',
+              itemHash: 'KFQrf4mEz0UPmUaFHwH4T6',
+              key: 'keyA0',
+            ),
+            'a0',
+          );
+        });
+
+        test('the complete item, when no key is given', () {
+          expect(
+            rljson.value(layer: '@layerA', itemHash: 'KFQrf4mEz0UPmUaFHwH4T6'),
+            {'keyA0': 'a0', '_hash': 'KFQrf4mEz0UPmUaFHwH4T6'},
+          );
+        });
+
+        test('the linked value, when property is a link', () {
+          final rljson = Rljson.exampleWithLink;
+
+          expect(
+            rljson.value(
+              layer: '@linkToLayerA',
+              itemHash: 'Cr5kvgDz5NGgbnuHu-Z05N',
+              key: '@layerA',
+            ),
+            {'_hash': 'KFQrf4mEz0UPmUaFHwH4T6', 'keyA0': 'a0'},
+          );
+        });
+
+        test('the linked value accross multiple layers using key2 to key4', () {
+          final rljson = Rljson.exampleWithDeepLink;
+          // print(const JsonEncoder.withIndent(' ')
+          // .convert(rljson.originalData));
+          // return;
+          expect(
+            rljson.value(
+              layer: '@a',
+              itemHash: 'GnSVp1CmoAo3rPiiGl44p-',
+              key: '@b',
+              key2: '@c',
+              key3: '@d',
+              key4: 'value',
+            ),
+            'd',
+          );
+        });
+      });
+
+      group('throws', () {
+        test('when key does not point to a valid value', () {
+          late final Exception exception;
+
+          try {
+            rljson.value(
+              layer: '@layerA',
+              itemHash: 'KFQrf4mEz0UPmUaFHwH4T6',
+              key: 'nonExistingKey',
+            );
+          } catch (e) {
+            exception = e as Exception;
+          }
+
+          expect(
+            exception.toString(),
+            'Exception: Key "nonExistingKey" not found in item with hash '
+            '"KFQrf4mEz0UPmUaFHwH4T6" in layer "@layerA"',
+          );
+        });
+
+        test('when a second key is given but the first key is not a link', () {
+          late final Exception exception;
+
+          try {
+            rljson.value(
+              layer: '@layerA',
+              itemHash: 'KFQrf4mEz0UPmUaFHwH4T6',
+              key: 'keyA0',
+              key2: 'keyA1',
+            );
+          } catch (e) {
+            exception = e as Exception;
+          }
+
+          expect(
+            exception.toString(),
+            'Exception: Invalid key "keyA1". '
+            'Additional keys are only allowed for links. '
+            'But key "keyA0" points to a value.',
+          );
         });
       });
     });